requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pymongo[srv]==3.12.0
# Remove the standalone bson package
sentence-transformers==2.2.2
python-multipart==0.0.6
python-dateutil==2.8.2
langchain-groq==0.3.2
langchain-core==0.3.52
langchain==0.3.23
numpy==1.24.3
python-dotenv==1.0.0
typing-extensions==4.8.0
gunicorn==21.2.0
torch==2.1.0
transformers==4.35.0
pydantic[email]